requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
confluent-kafka==2.10.0
distro==1.9.0
elevenlabs==2.1.0
fastapi==0.115.12
frozenlist==1.6.0
grpcio==1.72.1
grpcio-reflection==1.72.1
grpcio-tools==1.72.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.62
langchain-openai==0.3.18
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
langsmith==0.3.42
multidict==6.4.4
openai==1.82.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pillow==11.2.1
pluggy==1.6.0
prometheus_client==0.22.0
propcache==0.3.1
protobuf==6.31.1
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pytest==8.3.5
pytest-asyncio==1.0.0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==80.9.0
sniffio==1.3.1
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
